+++ src/components/MyFavorites.jsx
@@ -3,34 +3,56 @@ import { AuthContext } from './Provider/AuthProvider';
 import Loading from './Loading';
 import { fetchURL } from '../../fetchURL';
 import FavCard from './FavCard';
+import Swal from 'sweetalert2'
+
 
 const MyFavorites = () => {
-    const { userInfo, loading } = useContext(AuthContext);
+    const { user, loading } = useContext(AuthContext);
 
     const [favMovies, setFavMovies] = useState([]);
 
     const [countDelete, setcountDelete] = useState(1);
-    
-    useEffect (()=> {
-        fetch (`${fetchURL}/fav_movies/${userInfo?.email}`)
-        .then (res => res.json ())
-        .then (data => setFavMovies (data))
-    }, [userInfo, countDelete]);
 
-    console.log (favMovies);
+    useEffect(() => {
+        fetch(`${fetchURL}/fav_movies/${user?.email}`)
+            .then(res => res.json())
+            .then(data => setFavMovies(data))
+    }, [user, countDelete]);
 
     const handleDelete = (id) => {
 
-        fetch(`${fetchURL}/fav_movies/${id}`, {
-            method: 'DELETE'
-        })
-            .then(res => res.json())
-            .then(data => {
-                if( data.deletedCount>0) {
-                    setcountDelete (countDelete+1);
-                }
 
-            })
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`${fetchURL}/fav_movies/${id}`, {
+                    method: 'DELETE'
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        if (data.deletedCount > 0) {
+                            Swal.fire({
+                                title: "Deleted!",
+                                text: "Your file has been deleted.",
+                                icon: "success"
+                            });
+                            setcountDelete(countDelete + 1);
+                        }
+
+                    })
+
+
+
+            }
+        })
+
 
     }
 
@@ -42,7 +64,7 @@ const MyFavorites = () => {
 
             <div className='grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-10 place-items-center md:place-items-start'>
                 {
-                    favMovies.map ((movie, idx)=> <FavCard handleDelete ={handleDelete} key={idx} movie={movie}></FavCard>)
+                    favMovies.map((movie, idx) => <FavCard handleDelete={handleDelete} key={idx} movie={movie}></FavCard>)
                 }
             </div>
